examples/5_online_machine_learning/server_requirements.txt
svglib>=1.0.1
reportlab>=3.5.55
Flask>=1.1.1
opencv-python>=4.4.0.46
Pillow>=6.2.1
